+++ test/unit/server_usage/post_process.py
@@ -29,6 +29,43 @@ import version
 __version__ = version.__version__
 
 
+class ArgParser(object):
+
+    """Class:  ArgParser
+
+    Description:  Class stub holder for gen_class.ArgParser class.
+
+    Methods:
+        __init__
+        arg_exist
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.args_array = dict()
+
+    def arg_exist(self, arg):
+
+        """Method:  arg_exist
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_exist.
+
+        Arguments:
+
+        """
+
+        return True if arg in self.args_array else False
+
+
 class UnitTest(unittest.TestCase):
 
     """Class:  UnitTest
@@ -61,6 +98,13 @@ class UnitTest(unittest.TestCase):
         self.proc_data = {"Data": "String"}
         cfg = collections.namedtuple("Cfg", "db coll")
         self.cfg = cfg("database", "collection")
+        self.args = ArgParser()
+        self.args2 = ArgParser()
+        self.args3 = ArgParser()
+        self.args4 = ArgParser()
+        self.args.args_array = {"-n": True, "-m": True}
+        self.args2.args_array = {"-f": True}
+        self.args4.args_array = {"-n": True}
 
     @mock.patch("server_usage.mongo_libs.ins_doc")
     def test_mongo_failed(self, mock_mongo):
@@ -73,13 +117,12 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {"-n": True, "-m": True}
         mock_mongo.return_value = (False, "Connection Failure")
 
         with gen_libs.no_std_out():
             self.assertFalse(
                 server_usage.post_process(
-                    self.proc_data, args_array, self.cfg))
+                    self.proc_data, self.args, self.cfg))
 
     @mock.patch("server_usage.mongo_libs.ins_doc")
     def test_mongo_connect(self, mock_mongo):
@@ -92,11 +135,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {"-n": True, "-m": True}
         mock_mongo.return_value = (True, None)
 
         self.assertFalse(
-            server_usage.post_process(self.proc_data, args_array, self.cfg))
+            server_usage.post_process(self.proc_data, self.args, self.cfg))
 
     @mock.patch("server_usage.gen_libs.display_data")
     def test_print_format(self, mock_display):
@@ -109,11 +151,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {"-f": True}
         mock_display.return_value = True
 
         self.assertFalse(
-            server_usage.post_process(self.proc_data, args_array, self.cfg))
+            server_usage.post_process(self.proc_data, self.args2, self.cfg))
 
     def test_print_raw(self):
 
@@ -125,12 +166,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {}
-
         with gen_libs.no_std_out():
             self.assertFalse(
                 server_usage.post_process(
-                    self.proc_data, args_array, self.cfg))
+                    self.proc_data, self.args3, self.cfg))
 
     def test_suppress_false(self):
 
@@ -142,12 +181,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {}
-
         with gen_libs.no_std_out():
             self.assertFalse(
                 server_usage.post_process(
-                    self.proc_data, args_array, self.cfg))
+                    self.proc_data, self.args3, self.cfg))
 
     def test_suppress_true(self):
 
@@ -159,10 +196,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {"-n": True}
-
         self.assertFalse(
-            server_usage.post_process(self.proc_data, args_array, self.cfg))
+            server_usage.post_process(self.proc_data, self.args4, self.cfg))
 
     def test_mongo_false(self):
 
@@ -174,10 +209,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {"-n": True}
-
         self.assertFalse(
-            server_usage.post_process(self.proc_data, args_array, self.cfg))
+            server_usage.post_process(self.proc_data, self.args4, self.cfg))
 
     @mock.patch("server_usage.mongo_libs.ins_doc")
     def test_mongo_true(self, mock_mongo):
@@ -190,11 +223,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {"-n": True, "-m": True}
         mock_mongo.return_value = (True, None)
 
         self.assertFalse(
-            server_usage.post_process(self.proc_data, args_array, self.cfg))
+            server_usage.post_process(self.proc_data, self.args, self.cfg))
 
 
 if __name__ == "__main__":